Rename library name output and namespace of testing project
